requirements.txt
asgiref==3.8.1
asttokens==2.4.1
cfgv==3.4.0
decorator==5.1.1
distlib==0.3.9
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.1.0
filelock==3.16.1
identify==2.6.2
ipython==8.29.0
jedi==0.19.2
matplotlib-inline==0.1.7
nodeenv==1.9.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
pre_commit==4.0.1
prompt_toolkit==3.0.48
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.2
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
virtualenv==20.27.1
wcwidth==0.2.13
